Update pitch to int32_t

+++ src/OpenLoco/include/OpenLoco/Graphics/RenderTarget.h
@@ -15,8 +15,8 @@ namespace OpenLoco::Gfx
         int32_t y;          // 0x08
         int32_t width;      // 0x0C
         int32_t height;     // 0x10
-        int16_t pitch;      // 0x14 note: this is actually (pitch - width)
-        uint16_t zoomLevel; // 0x16
+        int32_t pitch;      // 0x14 note: this is actually (pitch - width)
+        uint16_t zoomLevel; // 0x18
 
         Ui::Rect getUiRect() const;
         Ui::Rect getDrawableRect() const;
